@@ -4,40 +4,41 @@
   "license": "Apache-2.0",
   "version": "1.0.0",
   "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "jest --detectOpenHandles -t 'Standard'",
-    "testSR25519": "jest --detectOpenHandles --runInBand tests/sr25519.test.js"
-  },
+  "types": "./dist/index.d.ts",
   "repository": {
     "type": "git",
     "url": "git+https://github.com/Zondax/ledger-polymesh"
   },
   "keywords": [
-    "zondax"
+    "Zondax",
+    "Ledger"
   ],
+  "scripts": {
+    "test": "jest --detectOpenHandles -t 'Standard'",
+    "testSR25519": "jest --detectOpenHandles --runInBand tests/sr25519.test.ts"
+  },
   "dependencies": {
-    "@zondax/ledger-polkadot": "^0.13.0",
-    "@zondax/zemu": "^0.11.7"
+    "@zondax/ledger-polkadot": "^0.13.6",
+    "@zondax/zemu": "^0.12.3"
   },
   "devDependencies": {
-    "@babel/cli": "^7.12.10",
-    "@babel/core": "^7.12.10",
-    "@babel/node": "^7.12.10",
-    "@babel/plugin-transform-runtime": "^7.12.10",
-    "@babel/preset-env": "^7.12.11",
-    "babel-eslint": "^10.1.0",
-    "babel-jest": "26.6.3",
+    "@types/jest": "^26.0.22",
+    "@types/ledgerhq__hw-transport": "^4.21.3",
+    "@typescript-eslint/eslint-plugin": "^4.20.0",
+    "@typescript-eslint/parser": "^4.20.0",
     "blakejs": "^1.1.0",
     "crypto-js": "4.0.0",
     "ed25519-supercop": "^2.0.1",
-    "eslint": "^7.8.1",
-    "eslint-config-prettier": "^7.1.0",
+    "eslint": "^7.24.0",
+    "eslint-config-prettier": "^8.2.0",
     "eslint-plugin-import": "^2.22.1",
-    "eslint-plugin-jest": "^24.1.3",
-    "eslint-plugin-prettier": "^3.3.1",
+    "eslint-plugin-jest": "^24.3.5",
+    "eslint-plugin-prettier": "^3.4.0",
     "jest": "26.6.3",
     "jest-serial-runner": "^1.1.0",
-    "jssha": "^3.1.1"
+    "jssha": "^3.1.1",
+    "prettier": "^2.0.4",
+    "ts-jest": "^26.5.4",
+    "typescript": "^4.2.3"
   }
 }
