@@ -14,35 +14,33 @@
  *  limitations under the License.
  ******************************************************************************* */
 
-import jest, {expect} from "jest";
-import Zemu from "@zondax/zemu";
+import Zemu, {DEFAULT_START_OPTIONS} from "@zondax/zemu";
+import {newPolymeshApp} from "@zondax/ledger-polkadot";
+
+// @ts-ignore
 import {blake2bFinal, blake2bInit, blake2bUpdate} from "blakejs";
-var addon = require('../../tests_tools/neon/native');
+import {APP_SEED} from "./common";
+const addon = require('../../tests_tools/neon/native');
 
-const {newPolymeshApp} = require("@zondax/ledger-polkadot");
 const Resolve = require("path").resolve;
+const APP_PATH = Resolve("../app/output/app_sr25519.elf");
 
-const APP_PATH_S = Resolve("../app/output/app_sr25519.elf");
-const APP_SEED = "equip will roof matter pink blind book anxiety banner elbow sun young"
-
-var simOptions = {
+const defaultOptions = {
+    ...DEFAULT_START_OPTIONS,
     logging: true,
-    start_delay: 3000,
     custom: `-s "${APP_SEED}"`,
-    X11: true
+    X11: false,
+    startDelay: 4000,
+    model: 'nanos'
 };
-
-let models = [
-    ['S', {model: 'nanos', prefix: 'S', path: APP_PATH_S}],
-]
 
 jest.setTimeout(60000)
 
 describe('SR25519', function () {
-    test.each(models)('get address (%s)', async function (_, {model, prefix, path}) {
-        const sim = new Zemu(path);
+    test('get address sr25519', async function () {
+        const sim = new Zemu(APP_PATH);
         try {
-            await sim.start(simOptions);
+            await sim.start({...defaultOptions});
             const app = newPolymeshApp(sim.getTransport());
 
             const resp = await app.getAddress(0x80000000, 0x80000000, 0x80000000, false, 1);
@@ -57,23 +55,22 @@ describe('SR25519', function () {
 
             expect(resp.address).toEqual(expected_address);
             expect(resp.pubKey).toEqual(expected_pk);
-
         } finally {
             await sim.close();
         }
     });
 
-    test.each(models)('show address (%s)', async function (_, {model, prefix, path}) {
-        const sim = new Zemu(path);
+    test('show address sr25519', async function () {
+        const sim = new Zemu(APP_PATH);
         try {
-            await sim.start(simOptions);
+            await sim.start({...defaultOptions});
             const app = newPolymeshApp(sim.getTransport());
 
             const respRequest = app.getAddress(0x80000000, 0x80000000, 0x80000000, true, 1);
             // Wait until we are not in the main menu
             await sim.waitUntilScreenIsNot(sim.getMainMenuSnapshot());
 
-            await sim.compareSnapshotsAndAccept(".", `${prefix.toLowerCase()}-show_address_sr25519`, 2);
+            await sim.compareSnapshotsAndAccept(".", "s-show_address_sr25519", 2);
 
             const resp = await respRequest;
             console.log(resp);
@@ -91,17 +88,16 @@ describe('SR25519', function () {
         }
     });
 
-    test.each(models)('show address - reject (%s)', async function (_, {model, prefix, path}) {
-        const sim = new Zemu(path);
+    test('show address - reject sr25519', async function () {
+        const sim = new Zemu(APP_PATH);
         try {
-            await sim.start(simOptions);
+            await sim.start({...defaultOptions});
             const app = newPolymeshApp(sim.getTransport());
 
             const respRequest = app.getAddress(0x80000000, 0x80000000, 0x80000000, true, 1);
             // Wait until we are not in the main menu
             await sim.waitUntilScreenIsNot(sim.getMainMenuSnapshot());
-
-            await sim.compareSnapshotsAndAccept(".", `${prefix.toLowerCase()}-show_address_reject_sr25519`, 3, 2);
+            await sim.compareSnapshotsAndAccept(".", "s-show_address_reject_sr25519", 3, 2);
 
             const resp = await respRequest;
             console.log(resp);
@@ -113,16 +109,16 @@ describe('SR25519', function () {
         }
     });
 
-    test.each(models)('sign basic normal (%s)', async function (_, {model, prefix, path}) {
-        const sim = new Zemu(path);
+    test('sign basic normal', async function () {
+        const sim = new Zemu(APP_PATH);
         try {
-            await sim.start(simOptions);
+            await sim.start({...defaultOptions});
             const app = newPolymeshApp(sim.getTransport());
             const pathAccount = 0x80000000;
             const pathChange = 0x80000000;
             const pathIndex = 0x80000000;
 
-            let txBlobStr = "040000287e8d895a3e642fa19a9c5ed97bec0d7cfc63bdb3641fc98a9c926cf0e8e93c0b63ce64c10c05d503910133158139ae28a3dfaac5fe1560a5e9e05ce1070000070000009deeb940c92ae02111c3bd5baca89970384f4c9849f02a1b2e53e66414d30f9f9deeb940c92ae02111c3bd5baca89970384f4c9849f02a1b2e53e66414d30f9f";
+            const txBlobStr = "0500ff40f53afc3e2ac603eae342a2f18e4cc456cd12e77aaf4eddea2045c4fb5d39966d0fd5030000e2070000070000009deeb940c92ae02111c3bd5baca89970384f4c9849f02a1b2e53e66414d30f9f9deeb940c92ae02111c3bd5baca89970384f4c9849f02a1b2e53e66414d30f9f";
 
             const txBlob = Buffer.from(txBlobStr, "hex");
 
@@ -134,9 +130,9 @@ describe('SR25519', function () {
             // Wait until we are not in the main menu
             await sim.waitUntilScreenIsNot(sim.getMainMenuSnapshot());
 
-            await sim.compareSnapshotsAndAccept(".", `${prefix.toLowerCase()}-sign_basic_normal_sr25519`, model === "nanos" ? 6 : 7);
+            await sim.compareSnapshotsAndAccept(".", "s-sign_basic_normal", 4);
 
-            let signatureResponse = await signatureRequest;
+            const signatureResponse = await signatureRequest;
             console.log(signatureResponse);
 
             expect(signatureResponse.return_code).toEqual(0x9000);
@@ -149,18 +145,18 @@ describe('SR25519', function () {
                 blake2bUpdate(context, txBlob);
                 prehash = Buffer.from(blake2bFinal(context));
             }
-            let signingcontext = Buffer.from([]);
-            const valid = addon.schnorrkel_verify(pubKey,signingcontext,prehash, signatureResponse.signature.slice(1));
+            const signingcontext = Buffer.from([]);
+            const valid = addon.schnorrkel_verify(pubKey, signingcontext, prehash, signatureResponse.signature.slice(1));
             expect(valid).toEqual(true);
         } finally {
             await sim.close();
         }
     });
 
-    test.each(models)('sign basic expert (%s)', async function (_, {model, prefix, path}) {
-        const sim = new Zemu(path);
+    test('sign basic expert', async function () {
+        const sim = new Zemu(APP_PATH);
         try {
-            await sim.start(simOptions);
+            await sim.start({...defaultOptions});
             const app = newPolymeshApp(sim.getTransport());
             const pathAccount = 0x80000000;
             const pathChange = 0x80000000;
@@ -171,11 +167,11 @@ describe('SR25519', function () {
             await sim.clickBoth();
             await sim.clickLeft();
 
-            let txBlobStr = "040000287e8d895a3e642fa19a9c5ed97bec0d7cfc63bdb3641fc98a9c926cf0e8e93c0b63ce64c10c05d503910133158139ae28a3dfaac5fe1560a5e9e05ce1070000070000009deeb940c92ae02111c3bd5baca89970384f4c9849f02a1b2e53e66414d30f9f9deeb940c92ae02111c3bd5baca89970384f4c9849f02a1b2e53e66414d30f9f";
+            const txBlobStr = "0500ff40f53afc3e2ac603eae342a2f18e4cc456cd12e77aaf4eddea2045c4fb5d39966d0fd5030000e2070000070000009deeb940c92ae02111c3bd5baca89970384f4c9849f02a1b2e53e66414d30f9f9deeb940c92ae02111c3bd5baca89970384f4c9849f02a1b2e53e66414d30f9f";
 
             const txBlob = Buffer.from(txBlobStr, "hex");
 
-            const responseAddr = await app.getAddress(pathAccount, pathChange, pathIndex, false , 1);
+            const responseAddr = await app.getAddress(pathAccount, pathChange, pathIndex, false, 1);
             const pubKey = Buffer.from(responseAddr.pubKey, "hex");
 
             // do not wait here.. we need to navigate
@@ -184,9 +180,9 @@ describe('SR25519', function () {
             // Wait until we are not in the main menu
             await sim.waitUntilScreenIsNot(sim.getMainMenuSnapshot());
 
-            await sim.compareSnapshotsAndAccept(".", `${prefix.toLowerCase()}-sign_basic_expert_sr25519`, model === "nanos" ? 12 : 13);
+            await sim.compareSnapshotsAndAccept(".", "s-sign_basic_expert", 10);
 
-            let signatureResponse = await signatureRequest;
+            const signatureResponse = await signatureRequest;
             console.log(signatureResponse);
 
             expect(signatureResponse.return_code).toEqual(0x9000);
@@ -199,24 +195,24 @@ describe('SR25519', function () {
                 blake2bUpdate(context, txBlob);
                 prehash = Buffer.from(blake2bFinal(context));
             }
-            let signingcontext = Buffer.from([]);
-            const valid = addon.schnorrkel_verify(pubKey,signingcontext,prehash, signatureResponse.signature.slice(1));
+            const signingcontext = Buffer.from([]);
+            const valid = addon.schnorrkel_verify(pubKey, signingcontext, prehash, signatureResponse.signature.slice(1));
             expect(valid).toEqual(true);
         } finally {
             await sim.close();
         }
     });
 
-    test.each(models)('sign large nomination (%s)', async function (_, {model, prefix, path}) {
-        const sim = new Zemu(path);
+    test('sign large nomination', async function () {
+        const sim = new Zemu(APP_PATH);
         try {
-            await sim.start(simOptions);
+            await sim.start({...defaultOptions});
             const app = newPolymeshApp(sim.getTransport());
             const pathAccount = 0x80000000;
             const pathChange = 0x80000000;
             const pathIndex = 0x80000000;
 
-            let txBlobStr = "080540006e34558d8a0255d837207a78038826d5d6d868803bf19b4ac79c1f25551b607e00ec437f7d7c48fdcc7a0ec4b81304509c2715e80c2eea6c037f12c8b6f82d721f004423c993381a2248ad4e932ad857a59885dcd7816511fc98f1508bc7f5a48629002898b4f6d6fe083f6519fbec6878b617bd78fa24374e17c890d0f9ed41a29871009009fbc535f95ad1de9c8a21a9065b6b75761c1a88e51185075a2cbf44d101420054f7a0a6bb84f3babb8a77d770f2f43f58dc5468bea686f94038cfc8084b6d7f0098e7890acb6a1f84116a7b551f92f4c172a965d390c9368b54d9163a1620677c009ed61e5fac1d7d5d3a7f49f493c193ed5a8638755d4aaa8ae962f5c1a137993600ecf83d0d9baef51d1e5b58b372994b021970074b703574c9b73c351279c6946f0052c204370e503d59a63f7309aeb42ba79ba21cab309d5356f2751178bca9fd7900304d53f51a0af0d280bcfc76ff307a1f6d8a5d4917bbe36cc157b3637f1e0f2200007e1bae1f1c9106734faec4c6ac2229bfa83d513d3c9edb14130ad825757b1c001603b70d5dd403242e01e428311746ed265c91135e4e2a0f74e5e520b741cc57000eea86b8cbf40083d66aff6631b6ac4c97ffe1dba7c04e67d6ba3ad0b9775648005ed646319954b783e397d2ae8dbf8d56e4868a7dbb57228d6c67be02b26712260008df51facad217e63431beeb63e42f7a3e03dfbedc1be252e83a61014f481e6ad503910100e1070000070000009deeb940c92ae02111c3bd5baca89970384f4c9849f02a1b2e53e66414d30f9f9deeb940c92ae02111c3bd5baca89970384f4c9849f02a1b2e53e66414d30f9f";
+            const txBlobStr = "110520ff3a74572ba8f3136c8fadd394cea53e7f1e8b6cb0dd35969db363abf52fc99b30ff305429035c83eab3af45c8324d5bc36a659f2cedfb36198156d0287f7f501107ff1a81532ef75c44acf3e5bf12d614b91fd5019276879bc8f0eac5a046b0c15277ffc8ee4c5e6d3ae0468077fa791e0468c90efff0ad9d3098b58f578d97a6465801ffc4faf3ff6cea6d0cc39dce4af2f126087a5b4221a0fe9903d22aaf5bd248e010ff06ea6ede1e2dd9cc9ab286762ed6732db47df39fc0f9d5a47a53325100c94f75fff065ea2f702c2cdd21abcfc6867404423fe2c0ecce2fecbdcfee28d481994d18ff8a9170375fef6774fe9381eace15328d1a2cce7809ced520edd75be93567e82dd5038d2400e2070000070000009deeb940c92ae02111c3bd5baca89970384f4c9849f02a1b2e53e66414d30f9f9deeb940c92ae02111c3bd5baca89970384f4c9849f02a1b2e53e66414d30f9f";
 
             const txBlob = Buffer.from(txBlobStr, "hex");
 
@@ -228,9 +224,9 @@ describe('SR25519', function () {
             // Wait until we are not in the main menu
             await sim.waitUntilScreenIsNot(sim.getMainMenuSnapshot());
 
-            await sim.compareSnapshotsAndAccept(".", `${prefix.toLowerCase()}-sign_large_nomination_sr25519`, model === "nanos" ? 33 : 18);
+            await sim.compareSnapshotsAndAccept(".", `s-sign_large_nomination_sr25519`, 17);
 
-            let signatureResponse = await signatureRequest;
+            const signatureResponse = await signatureRequest;
             console.log(signatureResponse);
 
             expect(signatureResponse.return_code).toEqual(0x9000);
@@ -243,12 +239,11 @@ describe('SR25519', function () {
                 blake2bUpdate(context, txBlob);
                 prehash = Buffer.from(blake2bFinal(context));
             }
-            let signingcontext = Buffer.from([]);
-            const valid = addon.schnorrkel_verify(pubKey,signingcontext,prehash, signatureResponse.signature.slice(1));
+            const signingcontext = Buffer.from([]);
+            const valid = addon.schnorrkel_verify(pubKey, signingcontext, prehash, signatureResponse.signature.slice(1));
             expect(valid).toEqual(true);
         } finally {
             await sim.close();
         }
     });
-
 });
