@@ -33,7 +33,7 @@ static char *secret_message =
         "USE AT YOUR OWN RISK!! "
         "You are about to enable the KSM recovery mode."
         "If you are not sure why you are here, reject or unplug your device immediately."
-        "Activating this mode will temporarily allow you to sign Kusama transactions using Polkadot keys";
+        "Activating this mode will temporarily allow you to sign transactions using Polkadot keys";
 
 zxerr_t secret_getNumItems(uint8_t *num_items) {
     zemu_log_stack("secret_getNumItems");
@@ -56,11 +56,9 @@ zxerr_t secret_getItem(int8_t displayIdx,
 
 zxerr_t secret_enabled() {
 #ifdef APP_SECRET_MODE_ENABLED
-#ifndef APP_RESTRICTED
-    zemu_log("KSM RECOVERY TRIGGERED");
+    zemu_log("RECOVERY TRIGGERED");
     view_review_init(secret_getItem, secret_getNumItems, secret_accept);
     view_review_show();
-#endif
 #endif
     return zxerr_ok;
 }
