@@ -78,7 +78,6 @@ parser_error_t _readFundingRoundName_V7(parser_context_t* c, pd_FundingRoundName
 parser_error_t _readFundraiserName_V7(parser_context_t* c, pd_FundraiserName_V7_t* v);
 parser_error_t _readGrandpaEquivocationProof_V7(parser_context_t* c, pd_GrandpaEquivocationProof_V7_t* v);
 parser_error_t _readIdentityId_V7(parser_context_t* c, pd_IdentityId_V7_t* v);
-parser_error_t _readInvestorUid_V7(parser_context_t* c, pd_InvestorUid_V7_t* v);
 parser_error_t _readInvestorZKProofData_V7(parser_context_t* c, pd_InvestorZKProofData_V7_t* v);
 parser_error_t _readKeyOwnerProof_V7(parser_context_t* c, pd_KeyOwnerProof_V7_t* v);
 parser_error_t _readKeyValue_V7(parser_context_t* c, pd_KeyValue_V7_t* v);
@@ -112,6 +111,7 @@ parser_error_t _readOptionPipDescription_V7(parser_context_t* c, pd_OptionPipDes
 parser_error_t _readOptionPortfolioNumber_V7(parser_context_t* c, pd_OptionPortfolioNumber_V7_t* v);
 parser_error_t _readOptionReceiptDetails_V7(parser_context_t* c, pd_OptionReceiptDetails_V7_t* v);
 parser_error_t _readOptionRecordDateSpec_V7(parser_context_t* c, pd_OptionRecordDateSpec_V7_t* v);
+parser_error_t _readOptionScope_V7(parser_context_t* c, pd_OptionScope_V7_t* v);
 parser_error_t _readOptionTargetIdentities_V7(parser_context_t* c, pd_OptionTargetIdentities_V7_t* v);
 parser_error_t _readOptionTax_V7(parser_context_t* c, pd_OptionTax_V7_t* v);
 parser_error_t _readOptionUrl_V7(parser_context_t* c, pd_OptionUrl_V7_t* v);
@@ -588,13 +588,6 @@ parser_error_t _toStringIdentityId_V7(
     uint8_t pageIdx,
     uint8_t* pageCount);
 
-parser_error_t _toStringInvestorUid_V7(
-    const pd_InvestorUid_V7_t* v,
-    char* outValue,
-    uint16_t outValueLen,
-    uint8_t pageIdx,
-    uint8_t* pageCount);
-
 parser_error_t _toStringInvestorZKProofData_V7(
     const pd_InvestorZKProofData_V7_t* v,
     char* outValue,
@@ -821,6 +814,13 @@ parser_error_t _toStringOptionReceiptDetails_V7(
 
 parser_error_t _toStringOptionRecordDateSpec_V7(
     const pd_OptionRecordDateSpec_V7_t* v,
+    char* outValue,
+    uint16_t outValueLen,
+    uint8_t pageIdx,
+    uint8_t* pageCount);
+
+parser_error_t _toStringOptionScope_V7(
+    const pd_OptionScope_V7_t* v,
     char* outValue,
     uint16_t outValueLen,
     uint8_t pageIdx,
